requirements.txt
psutil>=7.2.0
requests>=2.32.3
python-dotenv>=1.0.1
pytest>=9.0.2
